refactor: 공통 타입 및 인터페이스 별도 파일에 분리
